@@ -1,22 +1,21 @@
 ﻿using Xunit;
 
-namespace Core.Domain.Tests
+namespace Core.Domain.Tests;
+
+public class PlayerTests
 {
-    public class PlayerTests
+
+    [Fact]
+    void Given_Player_To_String_Should_Include_Number_And_Name()
     {
+        // Arrange
+        var player = new Player {Name = "Agnes", PlayerNumber = 50};
 
-        [Fact]
-        void Given_Player_To_String_Should_Include_Number_And_Name()
-        {
-            // Arrange
-            var player = new Player {Name = "Agnes", PlayerNumber = 50};
+        // Act
+        var playerAsString = player.ToString();
 
-            // Act
-            var playerAsString = player.ToString();
-
-            // Assert
-            Assert.Equal("Agnes-50", playerAsString);
-        }
-
+        // Assert
+        Assert.Equal("Agnes-50", playerAsString);
     }
+
 }
