@@ -1,11 +1,10 @@
-﻿namespace Core.Domain
+﻿namespace Core.Domain;
+
+public class Address
 {
-    public class Address
-    {
-        public string Street { get; set; }
+    public string Street { get; set; }
 
-        public int Number { get; set; }
+    public int Number { get; set; }
 
-        public string Extension { get; set; }
-    }
+    public string Extension { get; set; }
 }
