@@ -1,13 +1,12 @@
 ﻿using System.Collections.Generic;
 
-namespace Core.Domain
+namespace Core.Domain;
+
+public class Team
 {
-    public class Team
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public ICollection<Player> Players { get; set; }
-    }
+    public ICollection<Player> Players { get; set; }
 }
