@@ -1,11 +1,10 @@
 ﻿using System.Collections.Generic;
 using Core.Domain;
 
-namespace Core.DomainServices
+namespace Core.DomainServices;
+
+public interface ICoachRepository
 {
-    public interface ICoachRepository
-    {
-        IEnumerable<Coach> GetCoaches();
-        Coach GetById(int id);
-    }
+    IEnumerable<Coach> GetCoaches();
+    Coach GetById(int id);
 }
