@@ -4,14 +4,13 @@ using System.Linq;
 using System.Threading.Tasks;
 using Core.Domain;
 
-namespace Core.DomainServices
+namespace Core.DomainServices;
+
+public interface IGameRepository
 {
-    public interface IGameRepository
-    {
-        IQueryable<Game> GetAll();
-        IEnumerable<Game> GetAllHomeGames();
-        IEnumerable<Game> GetAllExternalGames();
-        IEnumerable<Game> Filter(Func<Game, bool> filterExpressie);
-        Task AddGame(Game newGame);
-    }
+    IQueryable<Game> GetAll();
+    IEnumerable<Game> GetAllHomeGames();
+    IEnumerable<Game> GetAllExternalGames();
+    IEnumerable<Game> Filter(Func<Game, bool> filterExpressie);
+    Task AddGame(Game newGame);
 }
