@@ -1,87 +1,86 @@
-namespace Infrastructure.Tests
+namespace Infrastructure.Tests;
+
+public class GameRepositoryTests
 {
-    public class GameRepositoryTests
-    {
-        //[Fact]
-        //public void Given_Games_HomeGames_Can_Be_Retrieved()
-        //{
-        //    // Arrange
-        //    var sut = new GameRepository() { Games = CreateGames() };
+    //[Fact]
+    //public void Given_Games_HomeGames_Can_Be_Retrieved()
+    //{
+    //    // Arrange
+    //    var sut = new GameRepository() { Games = CreateGames() };
 
-        //    // Act
-        //    var homeGames = sut.GetAllHomeGames();
+    //    // Act
+    //    var homeGames = sut.GetAllHomeGames();
 
-        //    // Assert
-        //    Assert.Equal(5, homeGames.Count());
-        //}
+    //    // Assert
+    //    Assert.Equal(5, homeGames.Count());
+    //}
 
-        //[Fact]
-        //public void Given_Games_ExternalGames_Can_Be_Retrieved()
-        //{
-        //    // Arrange
-        //    var sut = new GameRepository() { Games = CreateGames() };
+    //[Fact]
+    //public void Given_Games_ExternalGames_Can_Be_Retrieved()
+    //{
+    //    // Arrange
+    //    var sut = new GameRepository() { Games = CreateGames() };
 
-        //    // Act
-        //    var externalGames = sut.GetAllExternalGames();
+    //    // Act
+    //    var externalGames = sut.GetAllExternalGames();
 
-        //    // Assert
-        //    Assert.Equal(3, externalGames.Count());
-        //}
+    //    // Assert
+    //    Assert.Equal(3, externalGames.Count());
+    //}
 
-        //[Fact]
-        //public void Filter_Given_Games_ExternalGames_Can_Be_Retrieved()
-        //{
-        //    // Arrange
-        //    var sut = new GameRepository() { Games = CreateGames() };
+    //[Fact]
+    //public void Filter_Given_Games_ExternalGames_Can_Be_Retrieved()
+    //{
+    //    // Arrange
+    //    var sut = new GameRepository() { Games = CreateGames() };
 
-        //    // Act
-        //    var externalGames = sut.Filter(g => !g.IsHomeGame);
+    //    // Act
+    //    var externalGames = sut.Filter(g => !g.IsHomeGame);
 
-        //    // Assert
-        //    Assert.Equal(3, externalGames.Count());
-        //}
+    //    // Assert
+    //    Assert.Equal(3, externalGames.Count());
+    //}
 
-        //[Fact]
-        //public void Filter_Given_Games_HomeGames_From_Date_Can_Be_Retrieved()
-        //{
-        //    // Arrange
-        //    var sut = new GameRepository() { Games = CreateGames() };
+    //[Fact]
+    //public void Filter_Given_Games_HomeGames_From_Date_Can_Be_Retrieved()
+    //{
+    //    // Arrange
+    //    var sut = new GameRepository() { Games = CreateGames() };
 
-        //    // Act
-        //    var externalGames = sut.Filter(g => g.IsHomeGame && g.PlayTime > new DateTime(2020, 10, 07));
+    //    // Act
+    //    var externalGames = sut.Filter(g => g.IsHomeGame && g.PlayTime > new DateTime(2020, 10, 07));
 
-        //    // Assert
-        //    Assert.Equal(2, externalGames.Count());
-        //}
+    //    // Assert
+    //    Assert.Equal(2, externalGames.Count());
+    //}
 
-        //[Fact]
-        //public void FilterAsExtension_Given_Games_HomeGames_From_Date_Can_Be_Retrieved()
-        //{
-        //    // Arrange
-        //    var sut = new GameRepository() { Games = CreateGames() };
+    //[Fact]
+    //public void FilterAsExtension_Given_Games_HomeGames_From_Date_Can_Be_Retrieved()
+    //{
+    //    // Arrange
+    //    var sut = new GameRepository() { Games = CreateGames() };
 
-        //    // Act
-        //    var externalGames = sut.Games.FilterGames(g => g.IsHomeGame && g.PlayTime > new DateTime(2020, 10, 07));
+    //    // Act
+    //    var externalGames = sut.Games.FilterGames(g => g.IsHomeGame && g.PlayTime > new DateTime(2020, 10, 07));
 
-        //    // Assert
-        //    Assert.Equal(2, externalGames.Count());
-        //}
+    //    // Assert
+    //    Assert.Equal(2, externalGames.Count());
+    //}
 
 
-        //private List<Game> CreateGames()
-        //{
-        //    return new List<Game>
-        //    {
-        //        new Game(new DateTime(2020, 10, 01, 12, 00, 00), true),
-        //        new Game(new DateTime(2020, 10, 02, 12, 00, 00), false),
-        //        new Game(new DateTime(2020, 10, 03, 12, 00, 00), true),
-        //        new Game(new DateTime(2020, 10, 04, 12, 00, 00), false),
-        //        new Game(new DateTime(2020, 10, 05, 12, 00, 00), false),
-        //        new Game(new DateTime(2020, 10, 06, 12, 00, 00), true),
-        //        new Game(new DateTime(2020, 10, 07, 12, 00, 00), true),
-        //        new Game(new DateTime(2020, 10, 08, 12, 00, 00), true),
-        //    };
+    //private List<Game> CreateGames()
+    //{
+    //    return new List<Game>
+    //    {
+    //        new Game(new DateTime(2020, 10, 01, 12, 00, 00), true),
+    //        new Game(new DateTime(2020, 10, 02, 12, 00, 00), false),
+    //        new Game(new DateTime(2020, 10, 03, 12, 00, 00), true),
+    //        new Game(new DateTime(2020, 10, 04, 12, 00, 00), false),
+    //        new Game(new DateTime(2020, 10, 05, 12, 00, 00), false),
+    //        new Game(new DateTime(2020, 10, 06, 12, 00, 00), true),
+    //        new Game(new DateTime(2020, 10, 07, 12, 00, 00), true),
+    //        new Game(new DateTime(2020, 10, 08, 12, 00, 00), true),
+    //    };
 
-        //}
-    }
+    //}
 }
