@@ -6,66 +6,65 @@ using Core.Domain;
 using Core.DomainServices;
 using Microsoft.EntityFrameworkCore;
 
-namespace Infrastructure
+namespace Infrastructure;
+
+public class GameRepository : IGameRepository
 {
-    public class GameRepository : IGameRepository
+    private readonly GameDbContext _context;
+
+    public GameRepository(GameDbContext context)
     {
-        private readonly GameDbContext _context;
+        _context = context;
+    }
 
-        public GameRepository(GameDbContext context)
+    /// <summary>
+    /// Example of explicit loading
+    /// </summary>
+    public void PreLoad()
+    {
+        var firstGame = _context.Games.FirstOrDefault();
+
+        if (firstGame != null)
         {
-            _context = context;
+            _context.Entry(firstGame).Collection(game => game.Players).Load();
+            _context.Entry(firstGame).Reference(game => game.Coach).Load();
         }
+    }
 
-        /// <summary>
-        /// Example of explicit loading
-        /// </summary>
-        public void PreLoad()
+    public IQueryable<Game> GetAll()
+    {
+        return _context.Games.Include(g => g.Coach).Include(g => g.Opponent).
+            Include(g => g.Players).ThenInclude(p => p.CareTakers);
+    }
+
+    public IEnumerable<Game> GetAllHomeGames()
+    {
+        var games = from game in _context.Games
+            where game.IsHomeGame
+            select game;
+
+        return games.ToList();
+    }
+
+    public IEnumerable<Game> GetAllExternalGames()
+    {
+        return _context.Games.Where(g => !g.IsHomeGame);
+    }
+
+    public IEnumerable<Game> Filter(Func<Game, bool> filterExpressie)
+    {
+        foreach (var game in _context.Games)
         {
-            var firstGame = _context.Games.FirstOrDefault();
-
-            if (firstGame != null)
+            if (filterExpressie(game))
             {
-                _context.Entry(firstGame).Collection(game => game.Players).Load();
-                _context.Entry(firstGame).Reference(game => game.Coach).Load();
+                yield return game;
             }
         }
+    }
 
-        public IQueryable<Game> GetAll()
-        {
-            return _context.Games.Include(g => g.Coach).Include(g => g.Opponent).
-                Include(g => g.Players).ThenInclude(p => p.CareTakers);
-        }
-
-        public IEnumerable<Game> GetAllHomeGames()
-        {
-            var games = from game in _context.Games
-                where game.IsHomeGame
-                select game;
-
-            return games.ToList();
-        }
-
-        public IEnumerable<Game> GetAllExternalGames()
-        {
-            return _context.Games.Where(g => !g.IsHomeGame);
-        }
-
-        public IEnumerable<Game> Filter(Func<Game, bool> filterExpressie)
-        {
-            foreach (var game in _context.Games)
-            {
-                if (filterExpressie(game))
-                {
-                    yield return game;
-                }
-            }
-        }
-
-        public async Task AddGame(Game newGame)
-        {
-            _context.Games.Add(newGame);
-            await _context.SaveChangesAsync();
-        }
+    public async Task AddGame(Game newGame)
+    {
+        _context.Games.Add(newGame);
+        await _context.SaveChangesAsync();
     }
 }
