@@ -1,77 +1,76 @@
 ﻿using Microsoft.EntityFrameworkCore.Migrations;
 
-namespace Infrastructure.Migrations
+namespace Infrastructure.Migrations;
+
+public partial class RemovedJoinEntity : Migration
 {
-    public partial class RemovedJoinEntity : Migration
+    protected override void Up(MigrationBuilder migrationBuilder)
     {
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropTable(
-                name: "PlayerGames");
+        migrationBuilder.DropTable(
+            name: "PlayerGames");
 
-            migrationBuilder.CreateTable(
-                name: "GamePlayer",
-                columns: table => new
-                {
-                    GamesId = table.Column<int>(type: "int", nullable: false),
-                    PlayersId = table.Column<int>(type: "int", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_GamePlayer", x => new { x.GamesId, x.PlayersId });
-                    table.ForeignKey(
-                        name: "FK_GamePlayer_Games_GamesId",
-                        column: x => x.GamesId,
-                        principalTable: "Games",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                    table.ForeignKey(
-                        name: "FK_GamePlayer_Players_PlayersId",
-                        column: x => x.PlayersId,
-                        principalTable: "Players",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
+        migrationBuilder.CreateTable(
+            name: "GamePlayer",
+            columns: table => new
+            {
+                GamesId = table.Column<int>(type: "int", nullable: false),
+                PlayersId = table.Column<int>(type: "int", nullable: false)
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_GamePlayer", x => new { x.GamesId, x.PlayersId });
+                table.ForeignKey(
+                    name: "FK_GamePlayer_Games_GamesId",
+                    column: x => x.GamesId,
+                    principalTable: "Games",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+                table.ForeignKey(
+                    name: "FK_GamePlayer_Players_PlayersId",
+                    column: x => x.PlayersId,
+                    principalTable: "Players",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+            });
 
-            migrationBuilder.CreateIndex(
-                name: "IX_GamePlayer_PlayersId",
-                table: "GamePlayer",
-                column: "PlayersId");
-        }
+        migrationBuilder.CreateIndex(
+            name: "IX_GamePlayer_PlayersId",
+            table: "GamePlayer",
+            column: "PlayersId");
+    }
 
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropTable(
-                name: "GamePlayer");
+    protected override void Down(MigrationBuilder migrationBuilder)
+    {
+        migrationBuilder.DropTable(
+            name: "GamePlayer");
 
-            migrationBuilder.CreateTable(
-                name: "PlayerGames",
-                columns: table => new
-                {
-                    PlayerID = table.Column<int>(type: "int", nullable: false),
-                    GameID = table.Column<int>(type: "int", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_PlayerGames", x => new { x.PlayerID, x.GameID });
-                    table.ForeignKey(
-                        name: "FK_PlayerGames_Games_GameID",
-                        column: x => x.GameID,
-                        principalTable: "Games",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                    table.ForeignKey(
-                        name: "FK_PlayerGames_Players_PlayerID",
-                        column: x => x.PlayerID,
-                        principalTable: "Players",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
+        migrationBuilder.CreateTable(
+            name: "PlayerGames",
+            columns: table => new
+            {
+                PlayerID = table.Column<int>(type: "int", nullable: false),
+                GameID = table.Column<int>(type: "int", nullable: false)
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_PlayerGames", x => new { x.PlayerID, x.GameID });
+                table.ForeignKey(
+                    name: "FK_PlayerGames_Games_GameID",
+                    column: x => x.GameID,
+                    principalTable: "Games",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+                table.ForeignKey(
+                    name: "FK_PlayerGames_Players_PlayerID",
+                    column: x => x.PlayerID,
+                    principalTable: "Players",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+            });
 
-            migrationBuilder.CreateIndex(
-                name: "IX_PlayerGames_GameID",
-                table: "PlayerGames",
-                column: "GameID");
-        }
+        migrationBuilder.CreateIndex(
+            name: "IX_PlayerGames_GameID",
+            table: "PlayerGames",
+            column: "GameID");
     }
 }
