@@ -1,10 +1,9 @@
 ﻿using Core.DomainServices;
 
-namespace Infrastructure
-{
-    public class OpponentRepository : IOpponentRepository
-    {
-        private readonly GameDbContext _context;
+namespace Infrastructure;
 
-    }
+public class OpponentRepository : IOpponentRepository
+{
+    private readonly GameDbContext _context;
+
 }
