@@ -1,16 +1,15 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Portal.Models
+namespace Portal.Models;
+
+public class LoginModel
 {
-    public class LoginModel
-    {
-        [Required]
-        public string Name { get; set; }
+    [Required]
+    public string Name { get; set; }
 
-        [Required]
-        [UIHint("password")]
-        public string Password { get; set; }
+    [Required]
+    [UIHint("password")]
+    public string Password { get; set; }
 
-        public string ReturnUrl { get; set; } = "/";
-    }
+    public string ReturnUrl { get; set; } = "/";
 }
