@@ -24,9 +24,9 @@
         <tbody>
 
             @foreach (var game in @Model)
-            {
-                <partial model="game" name="_GameRow"/>
-                @*<tr>
+        {
+            <partial model="game" name="_GameRow"/>
+            @*<tr>
                     <td>@($"{@game.PlayTime:dd-MM-yyyy HH:mm}")</td>
                     <td>@($"{@game.DepartureTime:HH:mm}")</td>
                     <td>@game.IsHomeGame</td>
@@ -36,7 +36,7 @@
                     <td>@game.Opponent?.Name</td>
                    
                 </tr>*@
-            }
+        }
 
         </tbody>
 
